requirements.txt
Django==2.0
djangorestframework==3.8.2
pyotp==2.3.0
pytz==2020.1
